refactor: change from usize to u64

+++ src/bars.rs
@@ -34,8 +34,8 @@ fn new_time_bar(len: u64) -> ProgressBar {
 // region implementarion
 
 pub fn foreach_progress(
-    reporter: Receiver<u64>, num_retrieving: usize, num_increment: usize, quiet: bool,
-) -> usize {
+    reporter: Receiver<u64>, num_retrieving: u64, num_increment: u64, quiet: bool,
+) -> u64 {
     let mut updated = 0;
     let perc_bar = if quiet { None } else { Some(new_wide_bar(num_retrieving.to_u64())) };
     for _ in reporter.iter() {
@@ -51,7 +51,7 @@ pub fn foreach_progress(
     updated
 }
 
-pub fn wait_with_progress(millis: usize, message: &str) {
+pub fn wait_with_progress(millis: u64, message: &str) {
     if millis > 10 {
         let delta = millis.min(500).to_u64();
         let delay = Duration::from_millis(delta);
